Extract site picker contract to a file
Note: A lambda is passed to provide the site store because the typical
injection is not possible due to the requirement to unconditionally
initialize the contract during activity initialization.
